mcp-servers/notes/pyproject.toml
[project]
name = "mcp-notes"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = ["mcp>=1.0.0", "supabase>=2.0.0", "python-dotenv>=1.0.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["server"]
